+++ src/cxhalo.F90
@@ -35,7 +35,7 @@ use com_senga
 !     ARGUMENTS
 !     =========
 
-REAL(KIND=dp), INTENT(OUT)            :: bigarr(nxbigl:nxbigr,nybigl:nybigr,nzbigl:nzbigr)
+REAL(KIND=8), INTENT(OUT)            :: bigarr(nxbigl:nxbigr,nybigl:nybigr,nzbigl:nzbigr)
 INTEGER, INTENT(IN)                      :: jndexl
 INTEGER, INTENT(IN)                      :: jndexr
 INTEGER, INTENT(IN)                      :: kndexl
